moved type defs for reportstate to useReportState
